move jetbrains install logic to a central place since it is the same for each editor
